FEAT: modified modal

+++ frontend/hoodies/src/feature/anonymous/article/anonymousArticleDetail.js
@@ -161,25 +161,45 @@ const AnonymousArticleDetail = () => {
   };
 
   const deleteHandler = async (event) => {
-    const response = await deleteArticle(article._id);
-    if (response.statusCode === 200) {
-      history.push("/board/anonymous");
-    } else {
-      console.log("게시글 삭제 에러");
-    }
+    Swal.fire({
+      title: '삭제하시겠습니까?',
+      showCancelButton: true,
+      cancelButtonText: '취소',
+      confirmButtonText: '삭제',
+      reverseButtons: true,
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        const response = await deleteArticle(article._id);
+        if (response.statusCode === 200) {
+          history.push("/board/anonymous");
+        } else {
+          console.log("게시글 삭제 에러");
+        }
+      }
+    })
   };
 
   const deleteCommentHandler = async (commentId) => {
-    const response = await deleteComment(article._id, commentId);
-    if (response.statusCode === 200) {
-      const response1 = await fetchArticle(article._id);
-      setArticle(response1);
-      setComments(response1.comments);
-      setCommentsMap(anonymousWriter(response1.comments, response1.writer));
-      setArticleWriter(response1.writer);
-    } else {
-      console.log("댓글 삭제 에러");
-    }
+    Swal.fire({
+      title: '삭제하시겠습니까?',
+      showCancelButton: true,
+      cancelButtonText: '취소',
+      confirmButtonText: '삭제',
+      reverseButtons: true,
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        const response = await deleteComment(article._id, commentId);
+        if (response.statusCode === 200) {
+          const response1 = await fetchArticle(article._id);
+          setArticle(response1);
+          setComments(response1.comments);
+          setCommentsMap(anonymousWriter(response1.comments, response1.writer));
+          setArticleWriter(response1.writer);
+        } else {
+          console.log("댓글 삭제 에러");
+        }
+      }
+    })
   };
 
   const modifyCommentHandler = async (commentId, newContent) => {
@@ -257,22 +277,32 @@ const AnonymousArticleDetail = () => {
         timerProgressBar: true,
       });
     } else {
-      const response = await reportArticle(location.state);
-      if (response.statusCode === 200) {
-        const response1 = await fetchArticle(location.state);
-        setArticle(response1);
-        setComments(response1.comments);
-        setCommentsMap(anonymousWriter(response1.comments, response1.writer));
-        setArticleWriter(response1.writer);
-        Swal.fire({
-          title: "게시글이 신고되었습니다.",
-          icon: "success",
-          timer: 2000,
-          timerProgressBar: true,
-        });
-      } else {
-        console.log("신고 실패");
-      }
+      Swal.fire({
+        title: '신고하시겠습니까?',
+        showCancelButton: true,
+        cancelButtonText: '취소',
+        confirmButtonText: '신고',
+        reverseButtons: true,
+      }).then(async (result) => {
+        if (result.isConfirmed) {
+          const response = await reportArticle(location.state);
+          if (response.statusCode === 200) {
+            const response1 = await fetchArticle(location.state);
+            setArticle(response1);
+            setComments(response1.comments);
+            setCommentsMap(anonymousWriter(response1.comments, response1.writer));
+            setArticleWriter(response1.writer);
+            Swal.fire({
+              title: "게시글이 신고되었습니다.",
+              icon: "success",
+              timer: 2000,
+              timerProgressBar: true,
+            });
+          } else {
+            console.log("신고 실패");
+          }
+        }
+      })
     }
   };
 
@@ -285,22 +315,32 @@ const AnonymousArticleDetail = () => {
         timerProgressBar: true,
       });
     } else {
-      const response = await reportComment(article._id, comment._id);
-      if (response.statusCode === 200) {
-        const response1 = await fetchArticle(location.state);
-        setArticle(response1);
-        setComments(response1.comments);
-        setCommentsMap(anonymousWriter(response1.comments, response1.writer));
-        setArticleWriter(response1.writer);
-        Swal.fire({
-          title: "선택한 댓글이 신고되었습니다.",
-          icon: "success",
-          timer: 2000,
-          timerProgressBar: true,
-        });
-      } else {
-        console.log("댓글 신고 오류");
-      }
+      Swal.fire({
+        title: '신고하시겠습니까?',
+        showCancelButton: true,
+        cancelButtonText: '취소',
+        confirmButtonText: '신고',
+        reverseButtons: true,
+      }).then(async (result) => {
+        if (result.isConfirmed) {
+          const response = await reportComment(article._id, comment._id);
+          if (response.statusCode === 200) {
+            const response1 = await fetchArticle(location.state);
+            setArticle(response1);
+            setComments(response1.comments);
+            setCommentsMap(anonymousWriter(response1.comments, response1.writer));
+            setArticleWriter(response1.writer);
+            Swal.fire({
+              title: "선택한 댓글이 신고되었습니다.",
+              icon: "success",
+              timer: 2000,
+              timerProgressBar: true,
+            });
+          } else {
+            console.log("댓글 신고 오류");
+          }
+        }
+      })
     }
   };
